Extract addMemberInjection()
