+++ compiler/src/main/java/com/squareup/anvil/compiler/codegen/dagger/InjectConstructorFactoryGenerator.kt
@@ -9,11 +9,9 @@ import com.squareup.anvil.compiler.codegen.PrivateCodeGenerator
 import com.squareup.anvil.compiler.codegen.asArgumentList
 import com.squareup.anvil.compiler.codegen.injectConstructor
 import com.squareup.anvil.compiler.codegen.mapToParameter
-import com.squareup.anvil.compiler.daggerDoubleCheckFqNameString
 import com.squareup.anvil.compiler.injectFqName
 import com.squareup.anvil.compiler.internal.asClassName
 import com.squareup.anvil.compiler.internal.buildFile
-import com.squareup.anvil.compiler.internal.capitalize
 import com.squareup.anvil.compiler.internal.classesAndInnerClass
 import com.squareup.anvil.compiler.internal.generateClassName
 import com.squareup.anvil.compiler.internal.safePackageString
@@ -131,27 +129,10 @@ internal class InjectConstructorFactoryGenerator : PrivateCodeGenerator() {
               if (memberInjectParameters.isEmpty()) {
                 addStatement("return newInstance($newInstanceArgumentList)")
               } else {
-                addStatement("val instance = newInstance($newInstanceArgumentList)")
-
-                val memberInjectorClassName = "${clazz.generateClassName()}_MembersInjector"
-                val memberInjectorClass = ClassName(packageName, memberInjectorClassName)
-
-                memberInjectParameters.forEachIndexed { index, parameter ->
-
-                  val property = memberInjectProperties[index]
-                  val propertyName = property.nameAsSafeName.asString()
-                  val functionName = "inject${propertyName.capitalize()}"
-
-                  val param = when {
-                    parameter.isWrappedInProvider -> parameter.name
-                    parameter.isWrappedInLazy ->
-                      "$daggerDoubleCheckFqNameString.lazy(${parameter.name})"
-                    else -> parameter.name + ".get()"
-                  }
-
-                  addStatement("%T.$functionName(instance, $param)", memberInjectorClass)
-                }
-                addStatement("return instance")
+                val instanceName = "instance"
+                addStatement("val $instanceName = newInstance($newInstanceArgumentList)")
+                addMemberInjection(packageName, clazz, memberInjectParameters, memberInjectProperties, instanceName)
+                addStatement("return $instanceName")
               }
             }
             .build()
